neighbor error fixed. 22 error not solved

+++ Types.h
@@ -1,4 +1,8 @@
 typedef unsigned long long mbr_t;
 typedef unsigned long long cmbr_t;
-typedef float coord_t;
+
+// change in gh code's CUDA file as well
+// typedef float coord_t;
+typedef double coord_t;
+
 typedef short int poly_size_t;
